README CHANGES:  Changed Wording
Change made at line 727. Please see the arrow, <<<---, below. 

------------------------
TEXT SNIPPET:
### Pre-populated database(s)

Put the database file in the `www` directory and open the database like   <<<--- CHANGED "like" TO "as follows" for the sake of readability, at first I thought the word "like" was a typo. - Sandi 
(**both** database location and `createFromLocation` items are required):
-------- END TEXT SNIPPET ---------

- Sandi