README: Added Note Re: Location of Dbase
On line 735, I added a sentence to clarify the location to place the database file.  I think this is needed because I was confused for days trying to figure out an error - not realizing that things change if my HTML file sits in a child folder.  I think it will help if this is explicitly mentioned.